// src/app/generated/model/supportedMimeTypeDTO.ts
/**
 * document-management
 * 1000kit document management
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface SupportedMimeTypeDTO {
  creationDate?: string;
  creationUser?: string;
  modificationDate?: string;
  modificationUser?: string;
  version?: number;
  id?: string;
  description?: string;
  name?: string;
}